@@ -7,36 +7,23 @@ interface Repository
     /**
      * Retrieve settings entry for the given key.
      * The configured values of entry filter will be used to filter the settings entries.
-     *
-     * @param string|array $key
-     * @param mixed $default
-     * @return mixed
      */
-    public function get($key, $default = null);
+    public function get(string|array $key, mixed $default = null): mixed;
 
     /**
      * Store settings entry for the given key.
      * The configured values of entry filter will be used to filter the settings entries.
-     *
-     * @param string|array $key
-     * @param mixed $value
-     * @return void
      */
-    public function set($key, $value = null);
+    public function set(string|array $key, mixed $value = null): void;
 
     /**
      * Destroy the settings entry for the given key.
-     *
-     * @param string|array $key
-     * @return bool
      */
-    public function forget($key);
+    public function forget(string|array $key): void;
 
     /**
      * Retrieve all settings entry.
      * The configured values of entry filter will be used to filter the settings entries.
-     *
-     * @return array
      */
-    public function all();
+    public function all(): array;
 }
