@@ -3,15 +3,14 @@
 namespace Smartisan\Settings;
 
 use Settings;
+use Smartisan\Settings\Settings as SettingsManager;
 
 trait HasSettings
 {
     /**
      * Retrieve the settings manager instance for this model.
-     *
-     * @return \Smartisan\Settings\Settings
      */
-    public function settings()
+    public function settings(): SettingsManager
     {
         return Settings::for($this);
     }
