@@ -9,12 +9,10 @@ class SettingsServiceProvider extends ServiceProvider
 {
     /**
      * Register any package services.
-     *
-     * @return void
      */
     public function register(): void
     {
-        $this->mergeConfigFrom(__DIR__ . '/../config/settings.php', 'settings');
+        $this->mergeConfigFrom(__DIR__.'/../config/settings.php', 'settings');
 
         $this->app->bind('settings', function () {
             $settingsConfig = config(
@@ -33,18 +31,16 @@ class SettingsServiceProvider extends ServiceProvider
 
     /**
      * Boot registered package services.
-     *
-     * @return void
      */
     public function boot(): void
     {
         if ($this->app->runningInConsole()) {
             $this->publishes([
-                __DIR__ . '/../config/settings.php' => config_path('settings.php'),
+                __DIR__.'/../config/settings.php' => config_path('settings.php'),
             ], 'config');
 
             $this->publishes([
-                __DIR__ . '/../database/migrations/create_settings_table.php.stub' => database_path('migrations/' . date('Y_m_d_His', time()) . '_create_settings_table.php'),
+                __DIR__.'/../database/migrations/create_settings_table.php.stub' => database_path('migrations/'.date('Y_m_d_His', time()).'_create_settings_table.php'),
             ], 'migrations');
         }
     }
